+++ src/Kiota.Builder/CodeDOM/CodeTypeBase.cs
@@ -15,7 +15,8 @@ namespace Kiota.Builder {
         public bool ActionOf {get;set;} = false;
         public bool IsNullable {get;set;} = true;
         public CodeTypeCollectionKind CollectionKind {get;set;} = CodeTypeCollectionKind.None;
-
+        public bool IsCollection { get { return CollectionKind != CodeTypeCollectionKind.None; } }
+        public bool IsArray { get { return CollectionKind == CodeTypeCollectionKind.Array; } }
         public ChildType BaseClone<ChildType>(CodeTypeBase source) where ChildType : CodeTypeBase
         {
             ActionOf = source.ActionOf;