+++ src/Kiota.Builder/Writers/StringExtensions.cs
@@ -1,0 +1,5 @@
+namespace Kiota.Builder.Writers {
+    public static class StringExtensions {
+        public static string StripArraySuffix(this string original) => string.IsNullOrEmpty(original) ? original : original.TrimEnd(']').TrimEnd('[');
+    }
+}